doc: Improve the IVI Generator Tutorial
* Add the missing gifs
* Don't create bundles on mac
* Signature fix for the dbus chapter
* Link to this tutorial instead of the Climate Example in some places

+++ examples/ivicore/qface-tutorial/chapter6-own-backend/backend_dbus/instrumentclusterbackend.cpp
@@ -144,7 +144,7 @@ void InstrumentClusterBackend::fetchFuel()
     auto reply = m_client->asyncCall("fuel");
     auto watcher = new QDBusPendingCallWatcher(reply, this);
     connect(watcher, &QDBusPendingCallWatcher::finished, this, [this](QDBusPendingCallWatcher *watcher) {
-        QDBusPendingReply<int> reply = *watcher;
+        QDBusPendingReply<qreal> reply = *watcher;
         if (reply.isError()) {
             qCritical() << reply.error();
         } else {
